fix: add FeatureServiceImpl
